Revert "Updated PowerShell formatter to set additionalProperties"
This reverts commit 46d291a11f8e6dcb11cac1a449b1d481d6f2d630.
